#59: Added 'generate' command to CLI
