+++ examples/domainmodel/package.json
@@ -41,7 +41,7 @@
         "build": "tsc",
         "watch": "tsc -watch",
         "lint": "eslint src --ext ts",
-        "langium:generate": "langium && npm run copygrammar",
+        "langium:generate": "langium generate && npm run copygrammar",
         "copygrammar": "copyfiles -f ./src/language-server/generated/grammar.json ./out/language-server/generated"
     },
     "dependencies": {
