+++ packages/generator-langium/langium-template/package.json
@@ -40,7 +40,7 @@
         "build": "webpack",
         "watch": "webpack --watch",
         "lint": "eslint src --ext ts",
-        "langium:generate": "langium"
+        "langium:generate": "langium generate"
     },
     "dependencies": {
         "langium": "0.0.0",
