+++ packages/langium/package.json
@@ -22,7 +22,7 @@
     "watch": "tsc --watch",
     "lint": "eslint ./src --ext .js,.ts",
     "test": "jest --coverage",
-    "bootstrap": "node ../langium-cli/bin/langium && npm run copygrammar",
+    "bootstrap": "node ../langium-cli/bin/langium generate && npm run copygrammar",
     "copygrammar": "copyfiles -f ./src/grammar/generated/grammar.json ./lib/grammar/generated",
     "publish:next": "npm --no-git-tag-version version \"$(semver $npm_package_version -i minor)-next.$(git rev-parse --short HEAD)\" && npm publish --tag next",
     "publish:latest": "npm publish --tag latest"
